# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RevenuePoint {
  label: string;
  atRisk: number;
  recoverable: number;
  recovered: number;
}

export interface NamedCount {
  name: string;
  count: number;
  value: number;
}

export interface ChartData {
  revenue: RevenuePoint[];
  outcomes: NamedCount[];
  failureReasons: NamedCount[];
  actions: NamedCount[];
}

export interface Overview {
  revenueAtRisk: number;
  failedTransactions: number;
  potentiallyRecoverable: number;
  revenueRecovered: number;
  recoveryRate: number;
  successfulRecoveries: number;
  escalations: number;
  pendingManual: number;
  skipped: number;
  agentRun: boolean;
  charts: ChartData;
}

export interface Transaction {
  id: string;
  customerId: string;
  amount: number;
  currency: string;
  timestamp: string;
  paymentMethod: string;
  status: string;
  /** @nullable */
  failureReason: string | null;
  /** @nullable */
  recoveryScore: number | null;
  recoverableAmount: number;
  /** @nullable */
  recommendedAction: string | null;
  /** @nullable */
  guardrailDecision: string | null;
  /** @nullable */
  finalOutcome: string | null;
  /** @nullable */
  failureClassification: string | null;
  /** @nullable */
  decisionReason: string | null;
  retryCount: number;
}

export interface TransactionList {
  items: Transaction[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface AgentRunResult {
  overview: Overview;
  message: string;
  analyzed: number;
  completedAt: string;
}

export interface AuditEntry {
  id: string;
  timestamp: string;
  transactionId: string;
  aiScore: number;
  failureReason: string;
  recommendedAction: string;
  guardrailDecision: string;
  finalOutcome: string;
  reasoning: string;
}

export interface AuditList {
  items: AuditEntry[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface Guardrails {
  /**
     * @minimum 0
     * @maximum 10
     */
  maximumRetryAttempts: number;
  /** @minimum 0 */
  highValueThreshold: number;
  /**
     * @minimum 1
     * @maximum 1440
     */
  cooldownMinutes: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumRecoveryScore: number;
  /** @minimum 0 */
  maximumAutomaticRecoveryAmount: number;
}

export interface GuardrailsInput {
  /**
     * @minimum 0
     * @maximum 10
     */
  maximumRetryAttempts: number;
  /** @minimum 0 */
  highValueThreshold: number;
  /**
     * @minimum 1
     * @maximum 1440
     */
  cooldownMinutes: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumRecoveryScore: number;
  /** @minimum 0 */
  maximumAutomaticRecoveryAmount: number;
}

export type SearchParamParameter = string;

export type StatusParamParameter = string;

export type FailureReasonParamParameter = string;

export type OutcomeParamParameter = string;

export type PageParamParameter = number;

export type PageSizeParamParameter = number;

export type SortParamParameter = string;

export type SortDirectionParamParameter = typeof SortDirectionParamParameter[keyof typeof SortDirectionParamParameter];


export const SortDirectionParamParameter = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListTransactionsParams = {
search?: SearchParamParameter;
status?: StatusParamParameter;
failureReason?: FailureReasonParamParameter;
outcome?: OutcomeParamParameter;
/**
 * @minimum 1
 */
page?: PageParamParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParamParameter;
sort?: SortParamParameter;
sortDirection?: SortDirectionParamParameter;
};

export type ListAuditLogParams = {
search?: SearchParamParameter;
outcome?: OutcomeParamParameter;
/**
 * @minimum 1
 */
page?: PageParamParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParamParameter;
};