# lib/api-zod/src/generated/types/agentRunResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Overview } from './overview';

export interface AgentRunResult {
  overview: Overview;
  message: string;
  analyzed: number;
  completedAt: Date;
}