# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEntry {
  id: string;
  timestamp: Date;
  transactionId: string;
  aiScore: number;
  failureReason: string;
  recommendedAction: string;
  guardrailDecision: string;
  finalOutcome: string;
  reasoning: string;
}