# lib/api-zod/src/generated/types/chartData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NamedCount } from './namedCount';
import type { RevenuePoint } from './revenuePoint';

export interface ChartData {
  revenue: RevenuePoint[];
  outcomes: NamedCount[];
  failureReasons: NamedCount[];
  actions: NamedCount[];
}