# lib/api-zod/src/generated/types/guardrails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Guardrails {
  /**
     * @minimum 0
     * @maximum 10
     */
  maximumRetryAttempts: number;
  /** @minimum 0 */
  highValueThreshold: number;
  /**
     * @minimum 1
     * @maximum 1440
     */
  cooldownMinutes: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumRecoveryScore: number;
  /** @minimum 0 */
  maximumAutomaticRecoveryAmount: number;
}