# lib/api-zod/src/generated/types/overview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartData } from './chartData';

export interface Overview {
  revenueAtRisk: number;
  failedTransactions: number;
  potentiallyRecoverable: number;
  revenueRecovered: number;
  recoveryRate: number;
  successfulRecoveries: number;
  escalations: number;
  pendingManual: number;
  skipped: number;
  agentRun: boolean;
  charts: ChartData;
}