# lib/api-zod/src/generated/types/revenuePoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevenuePoint {
  label: string;
  atRisk: number;
  recoverable: number;
  recovered: number;
}