# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: string;
  customerId: string;
  amount: number;
  currency: string;
  timestamp: Date;
  paymentMethod: string;
  status: string;
  /** @nullable */
  failureReason: string | null;
  /** @nullable */
  recoveryScore: number | null;
  recoverableAmount: number;
  /** @nullable */
  recommendedAction: string | null;
  /** @nullable */
  guardrailDecision: string | null;
  /** @nullable */
  finalOutcome: string | null;
  /** @nullable */
  failureClassification: string | null;
  /** @nullable */
  decisionReason: string | null;
  retryCount: number;
}